Changed spine IDs for easy data passing to info div
I changed how the spines' IDs are set so that they are now the index of
the release in the JSON object. This makes passing the data to the info
div much easier (and a lot faster than searching for the ID). Next up
will be populating the info div with the rest of the release data.