Added Videos & prettyPhoto
Added div very similar to the track list div for video links, which
slides out from the left side of the info div. Since lightbox2 is not
able to show videos, I switched over to prettyPhoto.js, which is
significantly larger (and thus slower) but has almost the same
implementation (so it was easy to refactor the existing code to work
with it). Videos are now listed and play in a lightbox if clicked.