Left/right bounds to records (+multiple cover photos)
When scrolling, records will now stop when the first/last record is in
its 'selected' position (in the middle of the shelf).

Also, I forgot to note in my last commit that albums with multiple
images will display as a group in the lightbox (so the first image is
shown in the info div, and then others can be viewed in lightbox).