Added jQuery AJAX dataload
Since I haven't been able to get the records to display in any browser
except Firefox, I've decided to try switching the dataload over to
jQuery. So far, I have managed to get the data and parse it to JSON,
though it's not exactly the same as before. I'll keep working on it and
see what I can do.