Updated info to set tracks & videos on load
Moved the population of the tracklist and videolist to the build
function. The only thing that is set onclick is the cover, since that
has to pull from the site. I also added a few logging statements and
fixed some issues in the tracklist implementation. Next TODOs: Add
keyboard events back in, Make shelf % of page height/width with
overflow, Get spinewidth/height working again. Go Falcons.