Updated cover images to use URI's
Updated covers to use URI150's and link to full-res images so that cover
images show up on Chrome/Safari. I guess there's an issue with the
Discogs API that is preventing hotlinking to the full-res images, so
we'll see what happens. Worst case, I guess I can add an "onerror" to
the images that point to the URI150's. That would be a bit of work, but
probably a solid solution if I can get it working.