Added object prototype
After watching a Pluralsight course on structuring Javascript, I wanted
to practice a bit of what I learned, so I added a Record object
prototype and encapsulated a bunch of the functionality.

I'm not sure whether this is what caused a million errors, or if they
were there before and I didn't notice them, but in any case, they are
fixed. _.reduce() is still broken, but I commented it out for now.